backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
anywidget==0.9.21
asttokens==3.0.1
async-timeout==4.0.3
attrs==25.4.0
beautifulsoup4==4.14.3
blessings==1.7
blinker==1.9.0
bqplot==0.12.45
branca==0.8.2
cattrs==26.1.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
cryptography==46.0.5
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.2.1
distro==1.9.0
earthengine-api==1.7.19
ee==0.2
eerepr==0.1.2
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.129.0
filelock==3.24.2
filetype==1.2.0
Flask==3.1.3
flatbuffers==25.9.23
folium==0.20.0
fonttools==4.62.1
frozenlist==1.8.0
fsspec==2026.2.0
future==1.0.0
geemap==0.37.2
geocoder==1.38.1
geoserver-rest==2.10.0
google-api-core==2.30.1
google-api-python-client==2.193.0
google-auth==2.48.0
google-auth-httplib2==0.3.1
google-cloud-core==2.5.1
google-cloud-storage==3.10.1
google-crc32c==1.8.0
google-genai==1.63.0
google-resumable-media==2.8.2
googleapis-common-protos==1.73.1
greenlet==3.3.1
groq==0.37.1
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==0.36.2
idna==3.11
ipyevents==2.0.4
ipyfilechooser==0.6.0
ipyleaflet==0.20.0
ipython==8.39.0
ipywidgets==8.1.8
itsdangerous==2.2.0
jedi==0.19.2
jh2==5.0.10
Jinja2==3.1.6
jq==1.11.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyter-leaflet==0.20.0
jupyterlab_widgets==3.0.16
kiwisolver==1.5.0
langchain==1.2.10
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.13
langchain-google-genai==4.2.0
langchain-groq==1.1.2
langchain-huggingface==1.2.0
langchain-text-splitters==1.1.0
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.6
langsmith==0.7.3
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
multidict==6.7.1
mypy_extensions==1.1.0
narwhals==2.18.1
niquests==3.17.0
numpy==2.2.6
openmeteo_requests==1.7.5
openmeteo_sdk==1.25.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==2.3.3
parso==0.8.6
pillow==12.1.1
platformdirs==4.9.2
plotly==6.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.2
protobuf==6.33.6
psycopg2==2.9.11
psygnal==0.15.1
pure_eval==0.2.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.0
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
pyperclip==1.11.0
pyshp==3.0.3
python-box==7.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
qh3==1.5.6
ratelim==0.1.6
requests==2.32.5
requests-cache==1.3.0
requests-toolbelt==1.0.0
retry-requests==2.0.0
rsa==4.9.1
scooby==0.11.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.46
stack-data==0.6.3
starlette==0.52.1
tenacity==9.1.4
tokenizers==0.22.2
tqdm==4.67.3
traitlets==5.14.3
traittypes==0.2.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
url-normalize==2.2.1
urllib3==2.6.3
urllib3-future==2.16.900
uuid_utils==0.14.0
uvicorn==0.41.0
wassima==2.0.5
wcwidth==0.6.0
websockets==15.0.1
Werkzeug==3.1.6
widgetsnbextension==4.0.15
xmltodict==1.0.4
xxhash==3.6.0
xyzservices==2026.3.0
yarl==1.22.0